Added welcome message
Added welcome_() function to send RPL 001-005 after successfully registering. Added a call to it in a conditional in the PASS, NICK, USER and PONG commands
Added LUSERS command which can either be called directly and is always sent after registering successfully. Sends information about the server
Added MOTD command which can either be called directly and is always sent after registering successfully. Sends a message of the day which is stored in ressources/motd.txt
Added ressources/motd.txt